+++ extras/AFC.py
@@ -27,7 +27,7 @@ except: raise error(ERROR_STR.format(import_lib="AFC_utils", trace=traceback.for
 try: from extras.AFC_stats import AFCStats
 except: raise error(ERROR_STR.format(import_lib="AFC_stats", trace=traceback.format_exc()))
 
-AFC_VERSION="1.0.22"
+AFC_VERSION="1.0.25"
 
 # Class for holding different states so its clear what all valid states are
 class State:
@@ -112,7 +112,8 @@ class afc:
         self.extrude_factor     = 1.
 
         # Config get section
-        self.moonraker_port         = config.get("moonraker_port", None)             # Port to connect to when interacting with moonraker. Used when there are multiple moonraker/klipper instances on a single host
+        self.moonraker_port         = config.get("moonraker_port", 7125)             # Port to connect to when interacting with moonraker. Used when there are multiple moonraker/klipper instances on a single host
+        self.moonraker_host         = config.get("moonraker_host", "http://localhost")
         self.moonraker_connect_to   = config.get("moonraker_timeout", 30)
         self.unit_order_list        = config.get('unit_order_list','')
         self.VarFile                = config.get('VarFile','../printer_data/config/AFC/AFC.var')# Path to the variables file for AFC configuration.
@@ -288,11 +289,9 @@ class afc:
         enough time to start before AFC tries to connect. This fixes a race condition that can
         happen between klipper and moonraker when first starting up.
         """
-        moonraker_port = ""
-        if self.moonraker_port is not None: moonraker_port = ":{}".format(self.moonraker_port)
 
         try:
-            self.moonraker = AFC_moonraker( moonraker_port, self.logger )
+            self.moonraker = AFC_moonraker( self.moonraker_host, self.moonraker_port, self.logger )
             if not self.moonraker.wait_for_moonraker( toolhead=self.toolhead, timeout=self.moonraker_connect_to ):
                 return False
             self.spoolman = self.moonraker.get_spoolman_server()
@@ -494,6 +493,14 @@ class afc:
         try:
             if 'virtual' in self.bypass.name:
                 bypass_state = self.bypass.sensor_enabled
+
+                # Make sure lane is not loaded before enabling virtual bypass, force switch
+                # to disabled if a lane is loaded
+                if bypass_state and self.current is not None:
+                    self.logger.error(f"Cannot set virtual bypass, {self.current} is currently loaded.")
+                    self.bypass.sensor_enabled = False
+                    return False
+
                 # Update filament present to match enable button so it updates in guis
                 self.bypass.filament_present = bypass_state