+++ extras/AFC_lane.py
@@ -77,7 +77,7 @@ class AFCLane:
         self.extruder_temp      = None
         self.bed_temp           = None
         self.td1_data           = {}
-        self.runout_lane        = 'NONE'
+        self.runout_lane        = None
         self.status             = AFCLaneState.NONE
         # END TODO
 
@@ -559,8 +559,8 @@ class AFCLane:
                 self.weight = 1000 # Defaulting weight to 1000 upon load
             else:
                 if self.unit_obj.check_runout(self):
-                    # Checking to make sure runout_lane is set and does not equal 'NONE'
-                    if  self.runout_lane != 'NONE':
+                    # Checking to make sure runout_lane is set
+                    if self.runout_lane is not None:
                         self._perform_infinite_runout()
                     else:
                         self._perform_pause_runout()
@@ -651,8 +651,8 @@ class AFCLane:
                             self.logger.info(f"Cannot get TD-1 data for {self.name}, either toolhead is loaded or hub shows filament in path")
 
                 elif self.prep_state == False and self.name == self.afc.current and self.afc.function.is_printing() and self.load_state and self.status != AFCLaneState.EJECTING:
-                    # Checking to make sure runout_lane is set and does not equal 'NONE'
-                    if  self.runout_lane != 'NONE':
+                    # Checking to make sure runout_lane is set
+                    if self.runout_lane is not None:
                         self._perform_infinite_runout()
                     else:
                         self._perform_pause_runout()
@@ -1098,6 +1098,18 @@ class AFCLane:
         """
         if not self.load_state:
             self.afc.error.AFC_error("Lane:{} is not loaded, cannot set loaded to toolhead for this lane.".format(self.name), pause=False)
+            return
+
+        # Do not set lane as loaded if virtual bypass or normal bypass is enabled/triggered
+        if self.afc.bypass.sensor_enabled:
+            disable_msg = ""
+            msg = f"Cannot set {self.name} as loaded, "
+
+            if 'virtual' in self.afc.bypass.name:
+                msg += "virtual "
+                disable_msg = " and disable"
+            msg += f"bypass is enabled.\nPlease unload{disable_msg} before trying to set lanes as loaded."
+            self.logger.error(msg)
             return
 
         self.afc.function.unset_lane_loaded()