+++ extras/AFC_spool.py
@@ -225,7 +225,16 @@ class AFCSpool:
             if lane not in self.afc.lanes:
                 self.logger.info('{} Unknown'.format(lane))
                 return
+
             cur_lane = self.afc.lanes[lane]
+            # Check if spool id is already assigned to a different lane, don't assign to current lane if id
+            # is already assigned
+            if SpoolID != '':
+                SpoolID = int(SpoolID)
+                if cur_lane.spool_id != SpoolID and any( SpoolID == lane.spool_id for lane in self.afc.lanes.values()):
+                    self.logger.error(f"SpoolId {SpoolID} already assigned to a lane, cannot assign to {lane}.")
+                    return
+
             self.set_spoolID(cur_lane, SpoolID)
 
     def _get_filament_values( self, filament, field, default=None):
@@ -307,7 +316,7 @@ class AFCSpool:
             self.logger.info("No LANE Defined")
             return
 
-        runout = gcmd.get('RUNOUT', '')
+        runout = gcmd.get('RUNOUT', 'NONE')
         # Check to make sure runout does not equal lane
         if lane == runout:
             self.logger.error("Lane({}) and runout({}) cannot be the same".format(lane, runout))
@@ -322,7 +331,7 @@ class AFCSpool:
             return
 
         cur_lane = self.afc.lanes[lane]
-        cur_lane.runout_lane = runout
+        cur_lane.runout_lane = None if runout == 'NONE' else runout
         self.afc.save_vars()
 
     cmd_RESET_AFC_MAPPING_help = "Resets all lane mapping in AFC"